requirements.txt
python-dotenv==1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
httpx>=0.24.0
discord.py>=2.3.0 